Move new vtctld to a new package
I very quickly ran into circular imports trying to keep things
colocated.
